make DateTimeParameterEditor a tsx, remove useState
